backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
bcrypt>=4.0.0,<5
# Ajusta el driver según tu DATABASE_URL (ej. postgresql):
psycopg2-binary>=2.9.0
